Keep push(null) in close() for streams that have not received a response
This preserves the 'end' event for close(code) called before any response
has arrived (matching Node's pending-stream behaviour that the existing
'validates input and manages stream state' test covers), while still
ending the readable from _destroy once a body has started so 'end' is
suppressed for error rstCodes.

Move the new tests into their own file.
