+++ src/js/node/http2.ts
@@ -2534,6 +2534,13 @@ class Http2Stream extends Duplex {
         validateFunction(callback, "callback");
         this.once("close", callback);
       }
+      // Node only ends the readable here for a pending stream (id not yet assigned); ids are
+      // assigned eagerly here, so "no response yet" is the closest equivalent. Once the peer has
+      // started a body the readable must be ended by _destroy so 'end' is suppressed for rstCodes
+      // that synthesize an error.
+      if ((this[bunHTTP2StreamStatus] & StreamState.StreamResponded) === 0) {
+        this.push(null);
+      }
       const { ending } = this._writableState;
       if (!ending) {
         // If the writable side of the Http2Stream is still open, emit the
